Edit structure and code.
